move renderSelection method to sprite
