+++ src/editor/editor-state.ts
@@ -113,16 +113,4 @@ export class EditorState {
       color: "rgba(0, 0, 0, 0.25)",
     })
   }
-
-  renderSpriteSelection(renderer: Renderer, sprite: SpriteState) {
-    const { left, top, width, height, corners } = sprite.rect
-    const color = "rgb(96, 165, 250)"
-    const props = { left, top, width, height, color }
-
-    renderer.fillRect({ ...props, alpha: 0.25 })
-    renderer.strokeRect({ ...props, lineWidth: 2 })
-    for (const corner of corners) {
-      renderer.fillArc({ x: corner.x, y: corner.y, radius: 5, color })
-    }
-  }
 }
